20210707-deoksu-4 수업 중 머지

+++ src/com/kkpjj/sysostory/controller/BossAttController.java
@@ -48,8 +48,8 @@ public class BossAttController extends JPanel{
 		int ranAtt = (int) (Math.random()*10)+1 ;
 		
 		int nomalAtt = monsterDTO.getMonAtt(); //일반공격
-		int spAtt = monsterDTO.getSkillAtt();  //스킬공격계수
-		int damege = nomalAtt * ( 1 + spAtt); // 스킬 데미지
+		double spAtt = monsterDTO.getSkillAtt();  //스킬공격계수
+		int damege = (int) (nomalAtt * ( 1 + spAtt)); // 스킬 데미지
 		
 		int chrHp = chr.getChrHp();
